+++ Tests/SteamPS.Tests.ps1
@@ -16,7 +16,7 @@ Describe "Get-SteamServerInfo" {
 Describe "Test SteamCMD cmdlets" {
     BeforeAll {
         . "$env:BHModulePath\Private\Add-EnvPath.ps1"
-        Add-EnvPath -Path 'TestDrive:\Test\SteamCMD' -Container Session
+        Add-EnvPath -Path 'TestDrive:\Test\SteamCMD'
 
         Install-SteamCMD -InstallPath 'TestDrive:\Test' -Force
     }